move ensure_admin_exists decorator next to where it's used
It's only used in one place. No need to have more modules then necessary.
